feat(build): use tsdown for web component build script
