+++ packages/ibm-products/package.json
@@ -73,10 +73,6 @@
     "@ibm/telemetry-js-config-generator": "^2.0.1",
     "@percy/cli": "^1.31.0",
     "@percy/storybook": "^6.0.4",
-    "@rollup/plugin-babel": "^6.0.4",
-    "@rollup/plugin-commonjs": "^28.0.3",
-    "@rollup/plugin-node-resolve": "^16.0.1",
-    "@rollup/plugin-typescript": "^12.1.2",
     "@storybook/addon-docs": "^9.1.1",
     "@storybook/addon-links": "^9.1.1",
     "@storybook/csf": "^0.1.13",
@@ -101,7 +97,6 @@
     "react-dom": "^19.1.0",
     "rimraf": "^6.0.1",
     "rollup": "^4.35.0",
-    "rollup-plugin-strip-banner": "^3.1.0",
     "sass": "^1.85.1",
     "storybook": "^9.1.1",
     "tsdown": "^0.14.1",
